Experimental fix for the CoC discontinuity issue.
This is an experimental fix for an issue that was raised in the Post
Processing stack ("DoF doesn't handle alpha cutout very well").

https://github.com/Unity-Technologies/PostProcessing/issues/40

This fix includes two new options.

- Allow resample CoC: Previously the blending ratio between
blurred/unblurred image was calculated in a half-resolution pass, and it
introduces upsampling issues. When this option is enabled, the blending
ratio is calculated in the final pass by sampling the depth again. This
may introduce performance loss on bandwidth-limited platforms (PS4 and
mobiles). This option is effective to remove noisy seams thatt are
observed between blurred/unblurred areas.
- Predilate: When this option is enabled, the far blur field will be
slightly dilated. This is effective to hide black seams that are
introduced by the "allow resample CoC" option.